Add files via upload
This assignment will give us an opportunity to explore encapsulation, abstraction, containment, and parametric polymorphism in Java. In CS 5002, we identified a List as an abstract data type and implemented two versions that stored string data: a LinkedList backed by a ListNode, and an an ArrayList backed by a linear array. In this assignment, we implement these as generic classes and interfaces.